[fix] command service method to writer
