+++ qiin-be/src/main/java/com/beyond/qiin/domain/booking/service/command/ReservationCommandServiceImpl.java
@@ -263,21 +263,6 @@ public class ReservationCommandServiceImpl implements ReservationCommandService 
         reservationWriter.save(reservation);
     }
 
-    // 자원 상태 변경 시 예약 상태 변경
-    @Override
-    @Transactional
-    public void updateReservationsForAsset(final Long assetId, final Integer assetStatus) {
-        // 1 = UNAVAILABLE, 2 = MAINTENANCE
-        if (assetStatus != 1 && assetStatus != 2) return;
-        if (assetStatus == null) return;
-        // pending, approved, using 대상(0, 1, 2)
-        List<Reservation> reservations = reservationWriter.findFutureUsableReservations(assetId);
-
-        for (Reservation reservation : reservations) {
-            reservation.markUnavailable("자원 사용 불가 상태에 따른 자동 취소");
-        }
-    }
-
     // 하드 딜리트
     public void hardDeleteReservation(final Long reservationId) {
         reservationWriter.hardDelete(reservationId);
