+++ qiin-be/src/test/java/com/beyond/qiin/domain/booking/service/command/UpdateReservationForAssetTest.java
@@ -4,6 +4,8 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.beyond.qiin.domain.booking.entity.Reservation;
+import com.beyond.qiin.domain.booking.repository.ReservationJpaRepository;
+import com.beyond.qiin.domain.booking.support.ReservationReader;
 import com.beyond.qiin.domain.booking.support.ReservationWriter;
 import java.util.List;
 import org.junit.jupiter.api.Test;
@@ -16,10 +18,13 @@ import org.mockito.junit.jupiter.MockitoExtension;
 public class UpdateReservationForAssetTest {
 
     @InjectMocks
-    private ReservationCommandServiceImpl reservationCommandService;
+    private ReservationWriter reservationWriter;
 
     @Mock
-    private ReservationWriter reservationWriter;
+    private ReservationJpaRepository reservationJpaRepository;
+
+    @Mock
+    private ReservationReader reservationReader;
 
     @Mock
     private Reservation reservation1;
@@ -30,13 +35,16 @@ public class UpdateReservationForAssetTest {
     @Test
     void updateReservationsForAsset_unavailable_shouldCancelReservations() {
         Long assetId = 100L;
+        int assetStatus = 1; // UNAVAILABLE
 
-        when(reservationWriter.findFutureUsableReservations(assetId)).thenReturn(List.of(reservation1, reservation2));
+        // Writer 내부 호출 (repository) mock
+        when(reservationJpaRepository.findFutureUsableReservationsByAsset(assetId))
+                .thenReturn(List.of(reservation1, reservation2));
 
-        // when
-        reservationCommandService.updateReservationsForAsset(assetId, 1); // UNAVAILABLE
+        // when: 실제 로직 실행
+        reservationWriter.updateReservationsForAsset(assetId, assetStatus);
 
-        // then
+        // then: markUnavailable 호출되었는지 검증
         verify(reservation1).markUnavailable("자원 사용 불가 상태에 따른 자동 취소");
         verify(reservation2).markUnavailable("자원 사용 불가 상태에 따른 자동 취소");
     }
